Move and document CqlLibraryIdentifier, CqlLibraryVersion and CqlVersionedLibraryIdentifier
