+++ Cql/CoreTests/Infrastructure/CqlLibraryStringTests.cs
@@ -1,4 +1,5 @@
 using CqlSdkPrototype.Infrastructure;
+using Hl7.Cql.Runtime;
 
 namespace CoreTests.Infrastructure
 {
