+++ Cql/CqlSdkPrototype/Invocation/LibraryInvoker.cs
@@ -2,6 +2,7 @@
 using CqlSdkPrototype.Infrastructure;
 using CqlSdkPrototype.Invocation.Internal;
 using Hl7.Cql.Abstractions;
+using Hl7.Cql.Runtime;
 
 namespace CqlSdkPrototype.Invocation;
